+++ Project3/Data/Fine_perturbation_100_50000_off0.400000.txt
@@ -1,0 +1,80 @@
+1  0
+1.01013  0
+1.02025  0
+1.03038  0
+1.04051  0
+1.05063  0
+1.06076  0
+1.07089  0
+1.08101  0
+1.09114  0
+1.10127  0
+1.11139  0
+1.12152  0
+1.13165  0
+1.14177  0
+1.1519  0
+1.16203  0
+1.17215  0
+1.18228  0
+1.19241  0
+1.20253  1
+1.21266  1
+1.22278  0
+1.23291  4
+1.24304  45
+1.25316  100
+1.26329  100
+1.27342  100
+1.28354  100
+1.29367  100
+1.3038  100
+1.31392  100
+1.32405  100
+1.33418  100
+1.3443  100
+1.35443  100
+1.36456  100
+1.37468  100
+1.38481  100
+1.39494  100
+1.40506  100
+1.41519  100
+1.42532  100
+1.43544  100
+1.44557  100
+1.4557  100
+1.46582  100
+1.47595  100
+1.48608  100
+1.4962  100
+1.50633  100
+1.51646  100
+1.52658  36
+1.53671  13
+1.54684  3
+1.55696  1
+1.56709  2
+1.57722  0
+1.58734  1
+1.59747  1
+1.60759  0
+1.61772  0
+1.62785  0
+1.63797  0
+1.6481  0
+1.65823  0
+1.66835  0
+1.67848  0
+1.68861  0
+1.69873  0
+1.70886  0
+1.71899  0
+1.72911  0
+1.73924  0
+1.74937  0
+1.75949  0
+1.76962  0
+1.77975  0
+1.78987  0
+1.8  0
