+++ Project3/Data/RK4_perturbation_100_50000_off0.100000.txt
@@ -48,7 +48,7 @@
 1.29192  0
 1.31515  0
 1.33838  0
-1.36162  100
+1.36162  97
 1.38485  100
 1.40808  100
 1.43131  3
@@ -84,7 +84,7 @@
 2.12828  0
 2.15152  0
 2.17475  0
-2.19798  87
+2.19798  91
 2.22121  0
 2.24444  0
 2.26768  0
