+++ Project3/Data/RK4_perturbation_100_50000_off0.400000.txt
@@ -9,7 +9,7 @@
 0.385859  0
 0.409091  0
 0.432323  0
-0.455556  11
+0.455556  13
 0.478788  0
 0.50202  0
 0.525253  0
@@ -18,7 +18,7 @@
 0.594949  0
 0.618182  0
 0.641414  0
-0.664646  1
+0.664646  0
 0.687879  99
 0.711111  0
 0.734343  0
@@ -42,8 +42,8 @@
 1.15253  0
 1.17576  0
 1.19899  0
-1.22222  2
-1.24545  81
+1.22222  1
+1.24545  79
 1.26869  100
 1.29192  100
 1.31515  100
@@ -55,8 +55,8 @@
 1.45455  100
 1.47778  100
 1.50101  100
-1.52424  57
-1.54747  2
+1.52424  55
+1.54747  5
 1.57071  1
 1.59394  0
 1.61717  0
@@ -82,11 +82,11 @@
 2.08182  0
 2.10505  0
 2.12828  0
-2.15152  50
+2.15152  39
 2.17475  100
 2.19798  100
 2.22121  100
-2.24444  36
+2.24444  33
 2.26768  0
 2.29091  0
 2.31414  0
