+++ Project3/Data/RK4_perturbation_100_50000_off0.700000.txt
@@ -1,15 +1,15 @@
 0.2  0
-0.223232  10
+0.223232  16
 0.246465  0
 0.269697  3
 0.292929  0
 0.316162  0
-0.339394  3
+0.339394  2
 0.362626  0
 0.385859  0
 0.409091  0
-0.432323  98
-0.455556  91
+0.432323  95
+0.455556  96
 0.478788  0
 0.50202  0
 0.525253  0
@@ -20,7 +20,7 @@
 0.641414  100
 0.664646  100
 0.687879  100
-0.711111  7
+0.711111  6
 0.734343  0
 0.757576  0
 0.780808  0
@@ -34,11 +34,11 @@
 0.966667  0
 0.989899  0
 1.01313  0
-1.03636  0
+1.03636  1
 1.0596  0
-1.08283  1
+1.08283  0
 1.10606  4
-1.12929  22
+1.12929  26
 1.15253  100
 1.17576  100
 1.19899  100
@@ -59,9 +59,9 @@
 1.54747  100
 1.57071  100
 1.59394  100
-1.61717  99
-1.6404  16
-1.66364  6
+1.61717  100
+1.6404  9
+1.66364  3
 1.68687  0
 1.7101  0
 1.73333  0
